requirements.txt
opencv-python
scikit-image
numpy
PyYAML
Pillow
scipy
easyocr